Move tests into __tests__ folders
`i18n.js` and `validate.js` (plus its tests) were moved into the util
folder.
