+++ src/components/common/__tests__/Alert.test.js
@@ -10,7 +10,7 @@
 
 import React from "react";
 import { shallow } from "enzyme";
-import Alert from "./Alert";
+import Alert from "../Alert";
 
 it("renders an alert correctly", () => {
   const message = "test message";
