+++ src/components/common/__tests__/Footer.test.js
@@ -10,7 +10,7 @@
 
 import React from "react";
 import { shallow } from "enzyme";
-import Footer from "./Footer";
+import Footer from "../Footer";
 
 it("renders without crashing", () => {
   shallow(<Footer />);
