+++ src/components/list/__tests__/DomainList.test.js
@@ -10,8 +10,8 @@
 
 import React from "react";
 import { shallow } from "enzyme";
-import DomainList from "./DomainList";
-import api from "../../util/api";
+import DomainList from "../DomainList";
+import api from "../../../util/api";
 
 const domains = ["domain1.com", "domain2.com", "domain3.com"];
