+++ src/index.js
@@ -18,7 +18,7 @@ import "react-bootstrap-typeahead/css/Typeahead.css";
 import "./scss/style.css";
 import Full from "./containers/Full";
 import api from "./util/api";
-import { setupI18n } from "./i18n";
+import { setupI18n } from "./util/i18n";
 
 // Before rendering anything, check if there is a session cookie.
 // Note: the user could have an old session, so the first API call
