+++ src/util/i18n.js
@@ -12,7 +12,7 @@ import i18n from "i18next";
 import XHR from "i18next-xhr-backend";
 import LanguageDetector from "i18next-browser-languagedetector";
 import { reactI18nextModule } from "react-i18next";
-import config from "./config";
+import config from "../config";
 
 export function setupI18n() {
   i18n
